Merge Tar branch (last version january 10th 2021 at 11:59 p.m.)

See merge request alid/card-games-framework!4

+++ src/Model/Card.hpp
@@ -12,17 +12,17 @@ private:
     const int id;
 
 protected:
-    int value;
+    float value;
 
 public:
     // Ctor and Dtor :
-    Card(const string &_name, int _id, int _value);
+    Card(const string &_name, int _id, float _value);
     virtual ~Card();
 
     // Getters and Setters :
     int getId() const;
     string getName() const;
-    int getValue() const;
+    float getValue() const;
 
     virtual bool operator== (Card const & other) const;
     virtual bool operator> (Card const & other) const;
