+++ src/Model/ColoredCard.hpp
@@ -48,11 +48,11 @@ private:
     COLOR color;
 
 public:
-    ColoredCard(const string &_name, int id, int _value, COLOR _color);
+    ColoredCard(const string &_name, int id, float _value, COLOR _color);
     ~ColoredCard() override;
 
     COLOR getColor() const;
-    void setValue(int v);
+    void setValue(float v);
 
     virtual bool operator==(ColoredCard const & other) const;
     friend ostream& operator<<(ostream& out, const ColoredCard &fs);
